Move components to components dir
