+++ src/App.vue
@@ -11,7 +11,7 @@
 import firebase from 'firebase'
 import config from '../config/config'
 
-import board from './Board.vue'
+import board from './components/Board.vue'
 
 firebase.initializeApp(config.firebase)
